+++ examples/com/sailthru/client/UserExample.java
@@ -3,19 +3,19 @@ package com.sailthru.client;
 import com.sailthru.client.exceptions.ApiException;
 import com.sailthru.client.handler.response.JsonResponse;
 import com.sailthru.client.params.User;
+
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-public class SendExample {
+public class UserExample {
     public static void main(String[] args) {
         String apiKey = "****";
         String apiSecret = "****";
         SailthruClient client = new SailthruClient(apiKey, apiSecret);
         try {
-            Send send = new Send();
-            send.setTemplate("my-template");
-            send.setEmail("praj@sailthru.com");
+            String sailthruId = "4d371896cc0c1adbb079b8d0";
+            User user = new User(sailthruId);
 
             Map<String, Object> vars = new HashMap<String, Object>();
             vars.put("name", "Prajwal Tuladhar");
@@ -24,21 +24,16 @@ public class SendExample {
             addressVars.put("city", "Jackson Heights");
             addressVars.put("zip", "11372");
             vars.put("address", addressVars);
+            user.setVars(vars);
 
-            send.setVars(vars);
+            JsonResponse response1 = client.getUser(user); // GET
 
-            send.setScheduleTime("+10 hours");
-            Map<String, Object> options = new HashMap<String, Object>();
-            options.put("behalf_email", "user@example.com");
-            options.put("test", 1);
+            JsonResponse response2 = client.saveUser(user);
 
-            send.setOptions(options);
-
-            JsonResponse response = client.send(send);
-            if (response.isOK()) {
-                System.out.println(response.getResponse());
+            if (response1.isOK()) {
+                System.out.println(response1.getResponse());
             } else {
-                System.out.println(response.getResponse().get("error").toString());
+                System.out.println(response1.getResponse().get("error").toString());
             }
         } catch (ApiException e) {
             // handle exception
